Basic Tutorial Menu Interface
Included PSD as well. I was thinking having the tutorial come up as soon as
the game loads, then a user could skip it if he were familiar. As intuitive
as we make the GUI, kids with no bg in stuff like this might be lost. I'll
get up the main menu look later today. Also, I was thinking of having some
basic particle effects or clouds moving in the bg. Moving clouds could be
done by simply scrolling an a couple images if we want to avoid tasking the
CPU with generating anything. Oh, and what I'm designing is with the mindset
that it all has to be done by scratch.